Add HasLinkMixin to take care of our default link collection and the onsave and access logic.
